+++ x/coostaking/module.go
@@ -10,21 +10,24 @@ import (
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/spf13/cobra"
 
-	"github.com/babylonlabs-io/babylon/v4/x/coostaking/keeper"
-	"github.com/babylonlabs-io/babylon/v4/x/coostaking/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+
+	"github.com/babylonlabs-io/babylon/v4/x/coostaking/keeper"
+	"github.com/babylonlabs-io/babylon/v4/x/coostaking/types"
 )
 
 var (
 	_ appmodule.HasBeginBlocker = AppModule{}
 	_ appmodule.HasEndBlocker   = AppModule{}
-	_ appmodule.HasGenesis      = AppModule{}
 	_ appmodule.AppModule       = AppModule{}
 	_ appmodule.HasServices     = AppModule{}
 	_ module.AppModuleBasic     = AppModuleBasic{}
+	_ module.HasGenesisBasics   = AppModuleBasic{}
+	_ module.HasGenesis         = AppModuleBasic{}
 )
 
 // ----------------------------------------------------------------------------
@@ -34,10 +37,12 @@ var (
 // AppModuleBasic implements the AppModuleBasic interface that defines the independent methods a Cosmos SDK module needs to implement.
 type AppModuleBasic struct {
 	cdc codec.BinaryCodec
+
+	k keeper.Keeper
 }
 
-func NewAppModuleBasic(cdc codec.BinaryCodec) AppModuleBasic {
-	return AppModuleBasic{cdc: cdc}
+func NewAppModuleBasic(cdc codec.BinaryCodec, k keeper.Keeper) AppModuleBasic {
+	return AppModuleBasic{cdc: cdc, k: k}
 }
 
 // Name returns the name of the module as a string
@@ -47,15 +52,34 @@ func (AppModuleBasic) Name() string {
 
 // RegisterLegacyAminoCodec registers the amino codec for the module, which is used to marshal and unmarshal structs to/from []byte in order to persist them in the module's KVStore
 func (AppModuleBasic) RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
+	types.RegisterCodec(cdc)
 }
 
 // RegisterInterfaces registers a module's interface types and their concrete implementations as proto.Message
 func (a AppModuleBasic) RegisterInterfaces(reg cdctypes.InterfaceRegistry) {
+	types.RegisterInterfaces(reg)
+}
+
+// ExportGenesis implements module.HasGenesis.
+func (a AppModuleBasic) ExportGenesis(ctx sdk.Context, cdc codec.JSONCodec) json.RawMessage {
+	genState := types.DefaultGenesis()
+	return cdc.MustMarshalJSON(genState)
+}
+
+// InitGenesis implements module.HasGenesis.
+func (a AppModuleBasic) InitGenesis(ctx sdk.Context, cdc codec.JSONCodec, gs json.RawMessage) {
+	var genState types.GenesisState
+
+	cdc.MustUnmarshalJSON(gs, &genState)
+	err := a.k.InitGenesis(ctx, genState)
+	if err != nil {
+		panic(err)
+	}
 }
 
 // DefaultGenesis returns a default GenesisState for the module, marshalled to json.RawMessage. The default GenesisState need to be defined by the module developer and is primarily used for testing
 func (AppModuleBasic) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	return nil
+	return cdc.MustMarshalJSON(types.DefaultGenesis())
 }
 
 // ValidateGenesis used to validate the GenesisState, given in its json.RawMessage form
@@ -65,7 +89,10 @@ func (AppModuleBasic) ValidateGenesis(cdc codec.JSONCodec, config client.TxEncod
 
 // RegisterGRPCGatewayRoutes registers the gRPC Gateway routes for the module
 func (AppModuleBasic) RegisterGRPCGatewayRoutes(clientCtx client.Context, mux *runtime.ServeMux) {
-	types.RegisterQueryHandlerClient(context.Background(), mux, types.NewQueryClient(clientCtx)) //nolint:errcheck // either we propagate the error up the stack, or don't check here.
+	err := types.RegisterQueryHandlerClient(context.Background(), mux, types.NewQueryClient(clientCtx))
+	if err != nil {
+		panic(err)
+	}
 }
 
 // GetTxCmd returns the root Tx command for the module. The subcommands of this root command are used by end-users to generate new transactions containing messages defined in the module
@@ -85,8 +112,6 @@ func (AppModuleBasic) GetQueryCmd() *cobra.Command {
 // AppModule implements the AppModule interface that defines the inter-dependent methods that modules need to implement
 type AppModule struct {
 	AppModuleBasic
-
-	k keeper.Keeper
 }
 
 func NewAppModule(
@@ -94,46 +119,24 @@ func NewAppModule(
 	k keeper.Keeper,
 ) AppModule {
 	return AppModule{
-		AppModuleBasic: NewAppModuleBasic(cdc),
-		k:              k,
+		AppModuleBasic: NewAppModuleBasic(cdc, k),
 	}
 }
 
 // RegisterServices implements appmodule.HasServices.
-func (am AppModule) RegisterServices(grpc.ServiceRegistrar) error {
-	return nil
-}
-
-// DefaultGenesis implements appmodule.HasGenesis.
-// Subtle: this method shadows the method (AppModuleBasic).DefaultGenesis of AppModule.AppModuleBasic.
-func (am AppModule) DefaultGenesis(appmodule.GenesisTarget) error {
-	return nil
-}
-
-// ExportGenesis implements appmodule.HasGenesis.
-func (am AppModule) ExportGenesis(context.Context, appmodule.GenesisTarget) error {
-	return nil
-}
-
-// InitGenesis implements appmodule.HasGenesis.
-func (am AppModule) InitGenesis(context.Context, appmodule.GenesisSource) error {
-	return nil
-}
-
-// ValidateGenesis implements appmodule.HasGenesis.
-// Subtle: this method shadows the method (AppModuleBasic).ValidateGenesis of AppModule.AppModuleBasic.
-func (am AppModule) ValidateGenesis(appmodule.GenesisSource) error {
+func (am AppModule) RegisterServices(cfg grpc.ServiceRegistrar) error {
+	types.RegisterMsgServer(cfg, keeper.NewMsgServerImpl(am.k))
 	return nil
 }
 
 // BeginBlock implements appmodule.HasBeginBlocker.
-func (am AppModule) BeginBlock(context.Context) error {
-	return nil
+func (am AppModule) BeginBlock(ctx context.Context) error {
+	return BeginBlocker(ctx, am.k)
 }
 
 // EndBlock implements appmodule.HasEndBlocker.
-func (am AppModule) EndBlock(context.Context) error {
-	return nil
+func (am AppModule) EndBlock(ctx context.Context) error {
+	return EndBlocker(ctx, am.k)
 }
 
 // ConsensusVersion is a sequence number for state-breaking change of the module. It should be incremented on each consensus-breaking change introduced by the module. To avoid wrong/empty versions, the initial version should be set to 1